clock-sync/requirements.txt
bcrypt==4.1.2
grpcio==1.62.1
grpcio-tools==1.62.1
mysql-connector-python==8.3.0
protobuf==4.25.3
Pyro5==5.15
python-dateutil==2.9.0.post0
serpent==1.41
six==1.16.0
